refac: creating s3 service
