+++ import-service/package.json
@@ -15,7 +15,8 @@
     "@middy/http-json-body-parser": "^2.5.3",
     "aws-lambda": "^1.0.7",
     "aws-sdk": "^2.1183.0",
-    "csv-parser": "^3.0.0"
+    "csv-parser": "^3.0.0",
+    "dotenv": "^16.0.1"
   },
   "devDependencies": {
     "@serverless/typescript": "^3.21.0",
@@ -28,6 +29,7 @@
     "json-schema-to-ts": "^1.5.0",
     "serverless": "^3.0.0",
     "serverless-auto-swagger": "^2.8.2",
+    "serverless-dotenv-plugin": "^4.0.2",
     "serverless-esbuild": "^1.32.5",
     "serverless-offline": "^9.1.2",
     "ts-jest": "^28.0.7",
