+++ import-service/serverless.ts
@@ -3,12 +3,11 @@ import type { AWS } from '@serverless/typescript';
 import importFileParser from '@functions/importFileParser';
 import importProductsFile from '@functions/importProductsFile';
 
-const BUCKET = process.env.BUCKET;
-
 const serverlessConfiguration: AWS = {
   service: 'import-service',
   frameworkVersion: '3',
-  plugins: ['serverless-auto-swagger', 'serverless-esbuild', 'serverless-offline'],
+  useDotenv: true,
+  plugins: ['serverless-auto-swagger', 'serverless-esbuild', 'serverless-offline', 'serverless-dotenv-plugin'],
   provider: {
     name: 'aws',
     runtime: 'nodejs14.x',
@@ -19,17 +18,17 @@ const serverlessConfiguration: AWS = {
             {
               Effect: 'Allow',
               Action: 's3:ListBucket',
-              Resource: `arn:aws:s3:::${BUCKET}/*`
+              Resource: `arn:aws:s3:::${'${env:BUCKET}'}/*`
             },
             {
               Effect: 'Allow',
               Action: 's3:PutObject',
-              Resource: `arn:aws:s3:::${BUCKET}/uploaded/*`
+              Resource: `arn:aws:s3:::${'${env:BUCKET}'}/uploaded/*`
             },
             {
               Effect: 'Allow',
               Action: 's3:*',
-              Resource: `arn:aws:s3:::${BUCKET}/*`
+              Resource: `arn:aws:s3:::${'${env:BUCKET}'}/*`
             }
         ]
       }
