+++ import-service/src/functions/importFileParser/handler.ts
@@ -1,16 +1,12 @@
 import { APIGatewayProxyHandler } from 'aws-lambda';
-import { S3 } from 'aws-sdk';
-import csvParser from 'csv-parser';
+import { getAndParseBucketObject } from 'src/services/s3.service';
 import { middyfy } from '../../libs/lambda';
-
-const BUCKET = process.env.BUCKET;
 
 const importFileParser: APIGatewayProxyHandler = async (event) => {
   console.log(`event: ${JSON.stringify(event)}`);
-  const s3 = new S3({ region: 'us-east-1' });
 
   try {
-    await getAndParseBucketObject(event, s3);
+    await getAndParseBucketObject(event);
     return {
       statusCode: 200,
       body: JSON.stringify(''),
@@ -28,42 +24,3 @@ const importFileParser: APIGatewayProxyHandler = async (event) => {
 };
 
 export const main = middyfy(importFileParser);
-
-const getAndParseBucketObject = async (event, s3: S3) => {
-  for (const record of event.Records) {
-    console.log(record);
-    
-    const s3Stream = s3.getObject({
-      Bucket: BUCKET,
-      Key: record.s3.object.key
-    }).createReadStream();
-
-    const results = [];
-
-    s3Stream
-      .pipe(csvParser())
-      .on('data', (data) => results.push(data))
-      .on('error', (error) => {
-        console.log(error);
-      })
-      .on('end', () => {
-        results.map(item => {
-          console.log(`Adding item: ${JSON.stringify(item)}`);
-        })
-      });
-
-
-    await s3.copyObject({
-      Bucket: BUCKET,
-      CopySource: BUCKET + '/' + record.s3.object.key,
-      Key: record.s3.object.key.replace('uploaded', 'parsed')
-    }).promise();
-
-    await s3.deleteObject({
-      Bucket: BUCKET,
-      Key: record.s3.object.key,
-    }).promise()
-
-    console.log(`Excel file ` + record.s3.object.key.split('/')[1] + ' has been parsed!');
-  }
-}
