+++ import-service/src/functions/importFileParser/index.ts
@@ -1,13 +1,11 @@
 import { handlerPath } from '../../libs/handler-resolver';
 
-const BUCKET = process.env.BUCKET;
-
 export default {
   handler: `${handlerPath(__dirname)}/handler.main`,
   events: [
     {
       s3: {
-        bucket: BUCKET,
+        bucket: '${env:BUCKET}',
         event: 's3:ObjectCreated:*',
         existing: true,
         rules: [
